+++ changelog.md
@@ -7,14 +7,21 @@ __________          _________      ___________
  |____|_  /(____  /_______  / ____|/_______  /___|  /
         \/      \/        \/\/             \/     \/ 
 ```
+
 # RaSyEn Change Log
 
 An account of all that has happened and future plans.
+
+- **2.0.1**
+    - Added the =inline core filter.
+    - Fixed a bug related to `save-results` and joining lists.
+    - Tuned the `pronoun-swap`, and the `to-plural` en_US filter
 
 - **2.0.0**
     - Changed to two-digit versioning. Moving fast y'all :rocket:
     - Separated language related filters from main library.
     - Filters are language sorted (filters/en_US-filters.js, etc..)
-    - Added `pronoun` and `number-to-word` English related filters.
+    - Added `pronoun-swap` and `number-to-words` English related filters.
+
 - **< 1.9**
     - No change log record below 2.0, you may refer to git commit entries or rise an issue.